# docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: HaiTech CRM API
  description: |
    API for HaiTech CRM - מערכת לניהול הדרכות תכנות ובינה מלאכותית.
    
    ## Authentication
    - **JWT**: For user interface - `Authorization: Bearer <token>`
    - **API Key**: For integrations - `X-API-Key: <key>`
    
    ## Rate Limiting
    - Default: 1000 requests/hour
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
    
    ## Pagination
    - Use `limit` and `offset` query parameters
    - Default limit: 20, Max: 100
  version: 1.0.0
  contact:
    name: HaiTech Support
    email: support@hai.tech
  license:
    name: Proprietary
servers:
  - url: https://api.haitech-crm.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Development
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /meetings/{id}:
    get:
      tags: [Meetings]
      summary: Get meeting by ID
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Meeting details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MeetingFull'
                  meta:
                    $ref: '#/components/schemas/Meta'
    put:
      tags: [Meetings]
      summary: Update meeting
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingUpdate'
      responses:
        '200':
          description: Meeting updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Meta:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    Meeting:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cycleId:
          type: string
          format: uuid
        instructorId:
          type: string
          format: uuid
        scheduledDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        status:
          type: string
          enum: [scheduled, completed, cancelled, postponed]
        revenue:
          type: number
        instructorPayment:
          type: number
        profit:
          type: number
        topic:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    MeetingFull:
      allOf:
        - $ref: '#/components/schemas/Meeting'
        - type: object
          properties:
            cycle:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
            instructor:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
            attendance:
              type: array
              items:
                $ref: '#/components/schemas/Attendance'
    MeetingUpdate:
      type: object
      properties:
        instructorId:
          type: string
          format: uuid
        scheduledDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        topic:
          type: string
        notes:
          type: string

    # ==================
    # ATTENDANCE
    # ==================
    Attendance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        registrationId:
          type: string
          format: uuid
        status:
          type: string
          enum: [present, absent, late]
        isTrial:
          type: boolean
        notes:
          type: string
        student:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
...